# Goto binary serialization