The last commit is from https://github.com/model-checking/kani/pull/2221 which is needed for regression runs to pass (will be cleaned up on rebase).

### Description of changes: 

This adds a new output mode `--enable-unstable --write-goto-binary` that produces the symbol table a goto binary instead of a JSON file, and skips the invocation of `symtab2gb`.

The new serializer uses value numbering to share identical strings and nodes in the binary file.

This feature depends on this CBMC PR https://github.com/diffblue/cbmc/pull/7534, that makes it possible to load a goto binary produced by a different tool that may number strings differently from how CBMC numbers them.

### Resolved issues:

`Symtab2gb` causes some noticeable overhead.

- https://github.com/model-checking/kani/issues/2080
- https://github.com/model-checking/kani/issues/2074
- https://github.com/model-checking/kani/issues/2070

### Call-outs:

The build scripts still need to be updated so that the feature is tested as part of regression runs (help needed).

### Testing:

* How is this change tested?

- the module has internal tests and I ran the regression suite successfully with the option activated

* Is this a refactor change?

No

### Checklist
- [x] Each commit message has a non-empty body, explaining why the change was made
- [x] Methods or procedures are documented
- [x] Regression or unit tests are included, or existing tests cover the modified code
- [x] My PR is restricted to a single feature or bugfix

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 and MIT licenses.