@remi-delmas-3000 2023-02-10
> Thank you! Have you measured the performance improvement yet?

Serializing to GOTO binary seems to be as at least as fast or faster than when serializing to JSON, and we completely skip symtab2gb.

> I'm assuming you are planning to add a test as soon as the CBMC PR gets merged and released.

The module has its own internal tests that run at every build. 

As long as this is not used as default, would it be a good idea to have CI run the full regression suite a second time with GOTO binary output enabled (as done for the "unsound_experiments" feature) ?

I could add a couple of tests that specifically activate the switch but this would not be much compared to running the full regression suite.