@adpaco-aws 2023-02-10
> Serializing to GOTO binary seems to be as at least as fast or faster than when serializing to JSON, and we completely skip symtab2gb.

We would like to know about memory consumption as well. In fact, the biggest problem with `symtab2gb` (at least for Kani) is that launching several instances of it results in out-of-memory errors quite easily.