@remi-delmas-3000 2023-02-10
some performance metrics measured when running `scripts/std-lib-regression.sh` 
- total runtime goes from 197s to 70s
- peak memory consumption is reduced by two on the two largest crates `proc_macro` and `std` (see stats below)

# crate `std` :

## baseline
```
Command being timed: "/home/ubuntu/kani/target/kani/bin/kani-compiler --crate-name std --edition=2021 /home/ubuntu/.rustup/toolchains/nightly-2022-12-11-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/library/std/src/lib.rs --error-format=json --json=diagnostic-rendered-ansi,artifacts,future-incompat --crate-type rlib --emit=dep-info,metadata,link -C embed-bitcode=no -C debuginfo=2 --cfg feature="addr2line" --cfg feature="backtrace" --cfg feature="gimli-symbolize" --cfg feature="miniz_oxide" --cfg feature="object" --cfg feature="panic_unwind" --cfg feature="std_detect_dlsym_getauxval" --cfg feature="std_detect_file_io" -C metadata=4521dfe8186f7bad -C extra-filename=-4521dfe8186f7bad --out-dir /tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps --target x86_64-unknown-linux-gnu -Z force-unstable-if-unmarked -L dependency=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps -L dependency=/tmp/std_lib_test/target/debug/deps --extern addr2line=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libaddr2line-f4ba037b743064e3.rmeta --extern alloc=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/liballoc-d3e18ac105948c37.rmeta --extern cfg_if=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libcfg_if-fb37f524d97ee138.rmeta --extern compiler_builtins=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libcompiler_builtins-a5b7514ebb7379c4.rmeta --extern core=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libcore-672c5d05ab648bdb.rmeta --extern hashbrown=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libhashbrown-46445c7db4a1fb41.rmeta --extern libc=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/liblibc-7b17e0a66e9801f8.rmeta --extern miniz_oxide=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libminiz_oxide-b8e17be0cc86fe35.rmeta --extern object=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libobject-5c1de5f0d58e0ca9.rmeta --extern panic_abort=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libpanic_abort-82567b8412105d74.rmeta --extern panic_unwind=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libpanic_unwind-5ff4b67a1f33fae4.rmeta --extern rustc_demangle=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/librustc_demangle-15a62e4e79600337.rmeta --extern std_detect=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libstd_detect-2b7705bfffdedad6.rmeta --extern unwind=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libunwind-b20de0834f4fd475.rmeta --cap-lints allow --kani-compiler -Cpanic=abort -Cllvm-args=--goto-c -Cllvm-args=--ignore-global-asm -Cllvm-args=--reachability=legacy --cfg backtrace_in_libstd"
	User time (seconds): 78.80
	System time (seconds): 12.61
	Percent of CPU this job got: 99%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 1:31.43
	Average shared text size (kbytes): 0
	Average unshared data size (kbytes): 0
	Average stack size (kbytes): 0
	Average total size (kbytes): 0
	Maximum resident set size (kbytes): 6588252
	Average resident set size (kbytes): 0
	Major (requiring I/O) page faults: 0
	Minor (reclaiming a frame) page faults: 2293200
	Voluntary context switches: 7
	Involuntary context switches: 286
	Swaps: 0
	File system inputs: 0
	File system outputs: 1052376
	Socket messages sent: 0
	Socket messages received: 0
	Signals delivered: 0
	Page size (bytes): 4096
	Exit status: 0
    Finished dev [unoptimized + debuginfo] target(s) in 1m 50s
```

## goto binary output
```
write_goto_binary_file statistics GotoBinarySerializerStats { 
  allocated_bytes: 308575992, 
  string_stats: SharingStats { 
    nof_unique: 172792, 
    min_count: 1, 
    min_id: Some(3), 
    max_count: 450153, 
    max_id: Some(183), 
    avg_count: 22.182531598684793 
  }, 
  irep_stats: SharingStats { 
    nof_unique: 1068436,
    min_count: 1,
    min_id: Some(4),
    max_count: 144211,
    max_id: Some(14),
    avg_count: 3.2993693585764947 
  } 
}
Command being timed: "/home/ubuntu/kani/target/kani/bin/kani-compiler --crate-name std --edition=2021 /home/ubuntu/.rustup/toolchains/nightly-2022-12-11-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/library/std/src/lib.rs --error-format=json --json=diagnostic-rendered-ansi,artifacts,future-incompat --crate-type rlib --emit=dep-info,metadata,link -C embed-bitcode=no -C debuginfo=2 --cfg feature="addr2line" --cfg feature="backtrace" --cfg feature="gimli-symbolize" --cfg feature="miniz_oxide" --cfg feature="object" --cfg feature="panic_unwind" --cfg feature="std_detect_dlsym_getauxval" --cfg feature="std_detect_file_io" -C metadata=4521dfe8186f7bad -C extra-filename=-4521dfe8186f7bad --out-dir /tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps --target x86_64-unknown-linux-gnu -Z force-unstable-if-unmarked -L dependency=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps -L dependency=/tmp/std_lib_test/target/debug/deps --extern addr2line=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libaddr2line-f4ba037b743064e3.rmeta --extern alloc=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/liballoc-d3e18ac105948c37.rmeta --extern cfg_if=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libcfg_if-fb37f524d97ee138.rmeta --extern compiler_builtins=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libcompiler_builtins-a5b7514ebb7379c4.rmeta --extern core=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libcore-672c5d05ab648bdb.rmeta --extern hashbrown=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libhashbrown-46445c7db4a1fb41.rmeta --extern libc=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/liblibc-7b17e0a66e9801f8.rmeta --extern miniz_oxide=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libminiz_oxide-b8e17be0cc86fe35.rmeta --extern object=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libobject-5c1de5f0d58e0ca9.rmeta --extern panic_abort=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libpanic_abort-82567b8412105d74.rmeta --extern panic_unwind=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libpanic_unwind-5ff4b67a1f33fae4.rmeta --extern rustc_demangle=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/librustc_demangle-15a62e4e79600337.rmeta --extern std_detect=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libstd_detect-2b7705bfffdedad6.rmeta --extern unwind=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libunwind-b20de0834f4fd475.rmeta --cap-lints allow --kani-compiler -Cpanic=abort -Cllvm-args=--goto-c -Cllvm-args=--ignore-global-asm -Cllvm-args=--reachability=legacy --cfg backtrace_in_libstd"
	User time (seconds): 15.83
	System time (seconds): 9.66
	Percent of CPU this job got: 99%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 0:25.50
	Average shared text size (kbytes): 0
	Average unshared data size (kbytes): 0
	Average stack size (kbytes): 0
	Average total size (kbytes): 0
	Maximum resident set size (kbytes): 2759024
	Average resident set size (kbytes): 0
	Major (requiring I/O) page faults: 0
	Minor (reclaiming a frame) page faults: 1 006 544
	Voluntary context switches: 3
	Involuntary context switches: 75
	Swaps: 0
	File system inputs: 0
	File system outputs: 118224
	Socket messages sent: 0
	Socket messages received: 0
	Signals delivered: 0
	Page size (bytes): 4096
	Exit status: 0
    Finished dev [unoptimized + debuginfo] target(s) in 44.50s
```

# crate `proc_macro`
## baseline

```
Command being timed: "/home/ubuntu/kani/target/kani/bin/kani-compiler --crate-name proc_macro --edition=2021 /home/ubuntu/.rustup/toolchains/nightly-2022-12-11-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/library/proc_macro/src/lib.rs --error-format=json --json=diagnostic-rendered-ansi,artifacts,future-incompat --crate-type lib --emit=dep-info,metadata,link -C embed-bitcode=no -C debuginfo=2 -C metadata=e3dad344b94a38f1 -C extra-filename=-e3dad344b94a38f1 --out-dir /tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps --target x86_64-unknown-linux-gnu -Z force-unstable-if-unmarked -L dependency=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps -L dependency=/tmp/std_lib_test/target/debug/deps --extern core=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libcore-672c5d05ab648bdb.rmeta --extern std=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libstd-4521dfe8186f7bad.rmeta --cap-lints allow --kani-compiler -Cpanic=abort -Cllvm-args=--goto-c -Cllvm-args=--ignore-global-asm -Cllvm-args=--reachability=legacy"
	User time (seconds): 17.49
	System time (seconds): 3.58
	Percent of CPU this job got: 99%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 0:21.08
	Average shared text size (kbytes): 0
	Average unshared data size (kbytes): 0
	Average stack size (kbytes): 0
	Average total size (kbytes): 0
	Maximum resident set size (kbytes): 1 737 428
	Average resident set size (kbytes): 0
	Major (requiring I/O) page faults: 0
	Minor (reclaiming a frame) page faults: 610241
	Voluntary context switches: 6
	Involuntary context switches: 66
	Swaps: 0
	File system inputs: 0
	File system outputs: 284512
	Socket messages sent: 0
	Socket messages received: 0
	Signals delivered: 0
	Page size (bytes): 4096
	Exit status: 0
```

## with binary output
```
write_goto_binary_file statistics GotoBinarySerializerStats { 
  allocated_bytes: 76982280, 
  string_stats: SharingStats { 
    nof_unique: 49944, 
    min_count: 1, 
    min_id: Some(3), 
    max_count: 109757, 
    max_id: Some(177), 
    avg_count: 19.066154092583663 
  }, 
  irep_stats: SharingStats { 
    nof_unique: 269647,
    min_count: 1,
    min_id: Some(4),
    max_count: 39254,
    max_id: Some(14),
    avg_count: 3.2148327257488507 
  } 
}
Command being timed: "/home/ubuntu/kani/target/kani/bin/kani-compiler --crate-name proc_macro --edition=2021 /home/ubuntu/.rustup/toolchains/nightly-2022-12-11-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/library/proc_macro/src/lib.rs --error-format=json --json=diagnostic-rendered-ansi,artifacts,future-incompat --crate-type lib --emit=dep-info,metadata,link -C embed-bitcode=no -C debuginfo=2 -C metadata=e3dad344b94a38f1 -C extra-filename=-e3dad344b94a38f1 --out-dir /tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps --target x86_64-unknown-linux-gnu -Z force-unstable-if-unmarked -L dependency=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps -L dependency=/tmp/std_lib_test/target/debug/deps --extern core=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libcore-672c5d05ab648bdb.rmeta --extern std=/tmp/std_lib_test/target/x86_64-unknown-linux-gnu/debug/deps/libstd-4521dfe8186f7bad.rmeta --cap-lints allow --kani-compiler -Cpanic=abort -Cllvm-args=--goto-c -Cllvm-args=--ignore-global-asm -Cllvm-args=--reachability=legacy"
	User time (seconds): 3.90
	System time (seconds): 2.77
	Percent of CPU this job got: 99%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 0:06.68
	Average shared text size (kbytes): 0
	Average unshared data size (kbytes): 0
	Average stack size (kbytes): 0
	Average total size (kbytes): 0
	 Maximum resident set size (kbytes): 811320
    Maximum resident set size (kbytes): 1737428	
	Average resident set size (kbytes): 0
	Major (requiring I/O) page faults: 0
	Minor (reclaiming a frame) page faults: 280214
	Voluntary context switches: 2
	Involuntary context switches: 24
	Swaps: 0
	File system inputs: 0
	File system outputs: 28328
	Socket messages sent: 0
	Socket messages received: 0
	Signals delivered: 0
	Page size (bytes): 4096
	Exit status: 0
```